Alteracao dos nomes dos componentes para ingles
