--- src/components/pages/Reception/FinalizaPedido.js
@@ -1,13 +1,0 @@
-import React, {Component} from 'react';
-import CarrinhoCompra from 'components/ui/CarrinhoCompra';
-
-class FinalizaPedido extends Component {
-  
-  render() {
-    return (
-      <CarrinhoCompra></CarrinhoCompra>
-    )
-  }
-}  
-
-export default FinalizaPedido;
